Refactored the code by moving the ticket and screenshot structs into respective files in the models/ directory, and moving the CreateHoneyClient and GetHoneyClientById endpoint functions into the routes/ directory. Initialized a go.mod file so imports worked together, and the go.sum file is automatically generated similar to a package-lock.json file for npm.... we dont technically need this commit but wanted to abstract things out so we dont have merge conflicts in the future
