+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/csci4950tgt/api
+
+go 1.13
+
+require github.com/gorilla/mux v1.7.3
