+++ models/ticket.go
@@ -1,0 +1,8 @@
+package models
+
+type Ticket struct {
+	ID         int        `json:"id"`
+	Name       string     `json:"name"`
+	URL        string     `json:"url"`
+	Screenshot ScreenShot `json:"screenshot"`
+}
